requirements.txt
certifi==2023.5.7
charset-normalizer==3.2.0
filelock==3.12.2
idna==3.4
Jinja2==3.1.2
joblib==1.3.1
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.1
pandas==2.0.3
Pillow==10.0.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
sympy==1.12
threadpoolctl==3.1.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
